separation of 2 datasets
